+++ x/incentive/keeper/grpc_query_test.go
@@ -4,9 +4,14 @@ import (
 	"math/rand"
 	"testing"
 
+	"cosmossdk.io/collections"
+	"cosmossdk.io/math"
+	"cosmossdk.io/store/prefix"
+	storetypes "cosmossdk.io/store/types"
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	testkeeper "github.com/babylonlabs-io/babylon/testutil/keeper"
 	"github.com/babylonlabs-io/babylon/x/incentive/types"
+	"github.com/cosmos/cosmos-sdk/runtime"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 )
@@ -84,5 +89,65 @@ func FuzzBTCStakingGaugeQuery(f *testing.F) {
 			require.NoError(t, err)
 			require.True(t, resp.Gauge.Coins.Equal(gaugeList[i].Coins))
 		}
+	})
+}
+
+func FuzzDelegationRewardsQuery(f *testing.F) {
+	datagen.AddRandomSeedsToFuzzer(f, 10)
+	f.Fuzz(func(t *testing.T, seed int64) {
+		t.Parallel()
+		var (
+			r            = rand.New(rand.NewSource(seed))
+			storeKey     = storetypes.NewKVStoreKey(types.StoreKey)
+			k, ctx       = testkeeper.IncentiveKeeperWithStoreKey(t, storeKey, nil, nil, nil)
+			fp, del      = datagen.GenRandomAddress(), datagen.GenRandomAddress()
+			storeService = runtime.NewKVStoreService(storeKey)
+			store        = storeService.OpenKVStore(ctx)
+			storeAdaptor = runtime.KVStoreAdapter(store)
+			btcRwd       = datagen.GenRandomBTCDelegationRewardsTracker(r)
+		)
+
+		// Setup a BTCDelegationRewardsTracker for the delegator
+		require.NoError(t, k.BTCDelegationRewardsTracker.Set(ctx, collections.Join(fp.Bytes(), del.Bytes()), btcRwd))
+		st := prefix.NewStore(storeAdaptor, types.BTCDelegatorToFPKey)
+		delStore := prefix.NewStore(st, del.Bytes())
+		delStore.Set(fp.Bytes(), []byte{0x00})
+
+		// Setup FP current rewards with the corresponding period
+		// and same TotalActiveSat than the BTCDelegationRewardsTracker
+		fpCurrentRwd := datagen.GenRandomFinalityProviderCurrentRewards(r)
+		fpCurrentRwd.Period = btcRwd.StartPeriodCumulativeReward + 2
+		fpCurrentRwd.TotalActiveSat = btcRwd.TotalActiveSat
+		err := k.FinalityProviderCurrentRewards.Set(ctx, fp.Bytes(), fpCurrentRwd)
+		require.NoError(t, err)
+
+		// set start historical rewards corresponding to btcRwd.StartPeriodCumulativeReward
+		amtRwdInHistStart := fpCurrentRwd.CurrentRewards.MulInt(types.DecimalAccumulatedRewards).QuoInt(math.NewInt(2))
+		startHist := types.NewFinalityProviderHistoricalRewards(amtRwdInHistStart)
+		err = k.FinalityProviderHistoricalRewards.Set(ctx, collections.Join(fp.Bytes(), btcRwd.StartPeriodCumulativeReward), startHist)
+		require.NoError(t, err)
+
+		// set end period historical rewards
+		// end period for calculation is fpCurrentRwd.Period-1
+		amtRwdInHistEnd := amtRwdInHistStart.Add(fpCurrentRwd.CurrentRewards.MulInt(types.DecimalAccumulatedRewards).QuoInt(fpCurrentRwd.TotalActiveSat)...)
+		endHist := types.NewFinalityProviderHistoricalRewards(amtRwdInHistEnd)
+		err = k.FinalityProviderHistoricalRewards.Set(ctx, collections.Join(fp.Bytes(), fpCurrentRwd.Period-1), endHist)
+		require.NoError(t, err)
+
+		// Calculate expected rewards
+		expectedRwd := endHist.CumulativeRewardsPerSat.Sub(startHist.CumulativeRewardsPerSat...)
+		expectedRwd = expectedRwd.MulInt(btcRwd.TotalActiveSat.Add(fpCurrentRwd.TotalActiveSat))
+		expectedRwd = expectedRwd.QuoInt(types.DecimalAccumulatedRewards)
+
+		// Call the DelegationRewards query
+		res, err := k.DelegationRewards(
+			ctx,
+			&types.QueryDelegationRewardsRequest{
+				FinalityProviderAddress: fp.String(),
+				DelegatorAddress:        del.String(),
+			},
+		)
+		require.NoError(t, err)
+		require.Equal(t, expectedRwd, res.Rewards)
 	})
 }